+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Changes to this project are documented in this file. More detail (including info
 - AUTOTOOL-4278: Changing VirtualServer name causes changes on VirtualAddress properties
 - AUTOTOOL-4321: The requested Pool Member already exists in partition
 - AUTOTOOL-4319: Handle matching monitorNames properly
+- AUTOTOOL-4316: Add host, DeclarationID and tenant info to a failed declaration response
 
 ### Changed